@@ -18,16 +18,16 @@
     "lint": "lint-staged"
   },
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.14",
-    "@docusaurus/preset-classic": "2.0.0-beta.14",
-    "@docusaurus/theme-classic": "^3.0.0",
-    "@mdx-js/react": "^1.6.21",
+    "@docusaurus/core": "^3.1.1",
+    "@docusaurus/preset-classic": "^3.1.1",
+    "@docusaurus/theme-classic": "^3.1.1",
+    "@mdx-js/react": "^3.0.1",
     "clsx": "^1.1.1",
-    "mdx-mermaid": "^1.2.2",
-    "mermaid": "^8.14.0",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "mdx-mermaid": "^2.0.0",
+    "mermaid": "^10.8.0",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "browserslist": {
     "production": [
